+++ Net Pools/src/DynamicValueObjectPool.cs
@@ -152,25 +152,25 @@ namespace Enderlook.Pools
         /// Return rented object to pool.<br/>
         /// Default instances are discarded.
         /// </summary>
-        /// <param name="obj">Object to return.</param>
-        public override void Return(T obj)
+        /// <param name="element">Object to return.</param>
+        public override void Return(T element)
         {
-            if (EqualityComparer<T>.Default.Equals(obj, default))
+            if (EqualityComparer<T>.Default.Equals(element, default))
                 return;
 
             // Intentionally not using interlocked here.
             // In a worst case scenario two objects may be stored into same slot.
             // It is very unlikely to happen and will only mean that one of the objects will get collected.
-            if (firstElement.HasValue || !firstElement.TrySetValue(ref obj))
+            if (firstElement.HasValue || !firstElement.TrySetValue(ref element))
             {
                 ValueObjectWrapper<T>[] items = array;
                 for (int i = 0; i < items.Length; i++)
                 {
-                    if (!items[i].HasValue && items[i].TrySetValue(ref obj))
+                    if (!items[i].HasValue && items[i].TrySetValue(ref element))
                         break;
                 }
 
-                SendToReserve(obj);
+                SendToReserve(element);
             }
         }